refactor cluster implementation to have clear separation of raft/schema
