+++ cluster/cluster_service.go
@@ -17,17 +17,20 @@ import (
 	"time"
 
 	"github.com/sirupsen/logrus"
+	"github.com/weaviate/weaviate/cluster/bootstrap"
+	"github.com/weaviate/weaviate/cluster/resolver"
 	"github.com/weaviate/weaviate/cluster/rpc"
-	"github.com/weaviate/weaviate/cluster/store"
+	"github.com/weaviate/weaviate/cluster/schema"
 	enterrors "github.com/weaviate/weaviate/entities/errors"
 )
 
 // Service class serves as the primary entry point for the Raft layer, managing and coordinating
 // the key functionalities of the distributed consensus protocol.
 type Service struct {
-	*store.Service
+	*Raft
+
 	raftAddr string
-	config   *store.Config
+	config   *Config
 
 	rpcClient *rpc.Client
 	rpcServer *rpc.Server
@@ -38,18 +41,20 @@ type Service struct {
 	closeWaitForDB    chan struct{}
 }
 
-func New(cfg store.Config) *Service {
+// New returns a Service configured with cfg. The service will initialize internals gRPC api & clients to other cluster
+// nodes.
+// Raft store will be initialized and ready to be started. To start the service call Open().
+func New(cfg Config) *Service {
 	addr := fmt.Sprintf("%s:%d", cfg.Host, cfg.RPCPort)
-	cl := rpc.NewClient(rpc.NewResolver(cfg.IsLocalHost, cfg.RPCPort), cfg.RaftRPCMessageMaxSize)
-	fsm := store.New(cfg)
-	raftService := store.NewService(&fsm, cl)
+	cl := rpc.NewClient(resolver.NewRpc(cfg.IsLocalHost, cfg.RPCPort), cfg.RaftRPCMessageMaxSize)
+	fsm := NewFSM(cfg)
+	raft := NewRaft(&fsm, cl)
 	return &Service{
-		Service:  raftService,
-		raftAddr: fmt.Sprintf("%s:%d", cfg.Host, cfg.RaftPort),
-
+		Raft:              raft,
+		raftAddr:          fmt.Sprintf("%s:%d", cfg.Host, cfg.RaftPort),
 		config:            &cfg,
 		rpcClient:         cl,
-		rpcServer:         rpc.NewServer(&fsm, raftService, addr, cfg.Logger, cfg.RaftRPCMessageMaxSize),
+		rpcServer:         rpc.NewServer(&fsm, raft, addr, cfg.Logger, cfg.RaftRPCMessageMaxSize),
 		logger:            cfg.Logger,
 		closeBootstrapper: make(chan struct{}),
 		closeWaitForDB:    make(chan struct{}),
@@ -58,29 +63,29 @@ func New(cfg store.Config) *Service {
 
 // Open internal RPC service to handle node communication,
 // bootstrap the Raft node, and restore the database state
-func (c *Service) Open(ctx context.Context, db store.Indexer) error {
-	c.logger.WithField("servers", c.config.ServerName2PortMap).Info("open cluster service")
+func (c *Service) Open(ctx context.Context, db schema.Indexer) error {
+	c.logger.WithField("servers", c.config.NodeNameToPortMap).Info("open cluster service")
 	if err := c.rpcServer.Open(); err != nil {
 		return fmt.Errorf("start rpc service: %w", err)
 	}
 
-	if err := c.Service.Open(ctx, db); err != nil {
+	if err := c.Raft.Open(ctx, db); err != nil {
 		return fmt.Errorf("open raft store: %w", err)
 	}
 
-	bs := store.NewBootstrapper(
+	bs := bootstrap.NewBootstrapper(
 		c.rpcClient,
 		c.config.NodeID,
 		c.raftAddr,
-		c.config.AddrResolver,
-		c.Service.Ready,
+		c.config.NodeToAddressResolver,
+		c.Raft.Ready,
 	)
 
 	bCtx, bCancel := context.WithTimeout(ctx, c.config.BootstrapTimeout)
 	defer bCancel()
 	if err := bs.Do(
 		bCtx,
-		c.config.ServerName2PortMap,
+		c.config.NodeNameToPortMap,
 		c.logger,
 		c.config.Voter, c.closeBootstrapper); err != nil {
 		return fmt.Errorf("bootstrap: %w", err)
@@ -93,6 +98,8 @@ func (c *Service) Open(ctx context.Context, db store.Indexer) error {
 	return nil
 }
 
+// Close closes the raft service and frees all allocated ressources. Internal RAFT store will be closed and if
+// leadership is assumed it will be transferred to another node. gRPC server and clients will also be closed.
 func (c *Service) Close(ctx context.Context) error {
 	enterrors.GoWrapper(func() {
 		c.closeBootstrapper <- struct{}{}
@@ -100,7 +107,7 @@ func (c *Service) Close(ctx context.Context) error {
 	}, c.logger)
 
 	c.logger.Info("closing raft FSM store ...")
-	if err := c.Service.Close(ctx); err != nil {
+	if err := c.Raft.Close(ctx); err != nil {
 		return err
 	}
 
@@ -114,12 +121,13 @@ func (c *Service) Close(ctx context.Context) error {
 	return nil
 }
 
+// Ready returns or not whether the node is ready to accept requests.
 func (c *Service) Ready() bool {
-	return c.Service.Ready()
+	return c.Raft.Ready()
 }
 
 // LeaderWithID is used to return the current leader address and ID of the cluster.
 // It may return empty strings if there is no current leader or the leader is unknown.
 func (c *Service) LeaderWithID() (string, string) {
-	return c.Service.LeaderWithID()
+	return c.Raft.LeaderWithID()
 }
