+++ cluster/log.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package store
+package cluster
 
 import (
 	"errors"
