+++ cluster/raft_schema_endpoints.go
@@ -1,0 +1,204 @@
+//                           _       _
+// __      _____  __ ___   ___  __ _| |_ ___
+// \ \ /\ / / _ \/ _` \ \ / / |/ _` | __/ _ \
+//  \ V  V /  __/ (_| |\ V /| | (_| | ||  __/
+//   \_/\_/ \___|\__,_| \_/ |_|\__,_|\__\___|
+//
+//  Copyright © 2016 - 2024 Weaviate B.V. All rights reserved.
+//
+//  CONTACT: hello@weaviate.io
+//
+
+package cluster
+
+import (
+	"encoding/json"
+	"fmt"
+
+	"github.com/prometheus/client_golang/prometheus"
+	cmd "github.com/weaviate/weaviate/cluster/proto/api"
+	"github.com/weaviate/weaviate/cluster/schema"
+	"github.com/weaviate/weaviate/cluster/types"
+	"github.com/weaviate/weaviate/entities/models"
+	"github.com/weaviate/weaviate/usecases/monitoring"
+	"github.com/weaviate/weaviate/usecases/sharding"
+	"google.golang.org/protobuf/proto"
+)
+
+func (s *Raft) AddClass(cls *models.Class, ss *sharding.State) (uint64, error) {
+	if cls == nil || cls.Class == "" {
+		return 0, fmt.Errorf("nil class or empty class name : %w", schema.ErrBadRequest)
+	}
+
+	req := cmd.AddClassRequest{Class: cls, State: ss}
+	subCommand, err := json.Marshal(&req)
+	if err != nil {
+		return 0, fmt.Errorf("marshal request: %w", err)
+	}
+	command := &cmd.ApplyRequest{
+		Type:       cmd.ApplyRequest_TYPE_ADD_CLASS,
+		Class:      cls.Class,
+		SubCommand: subCommand,
+	}
+	return s.Execute(command)
+}
+
+func (s *Raft) UpdateClass(cls *models.Class, ss *sharding.State) (uint64, error) {
+	if cls == nil || cls.Class == "" {
+		return 0, fmt.Errorf("nil class or empty class name : %w", schema.ErrBadRequest)
+	}
+	req := cmd.UpdateClassRequest{Class: cls, State: ss}
+	subCommand, err := json.Marshal(&req)
+	if err != nil {
+		return 0, fmt.Errorf("marshal request: %w", err)
+	}
+	command := &cmd.ApplyRequest{
+		Type:       cmd.ApplyRequest_TYPE_UPDATE_CLASS,
+		Class:      cls.Class,
+		SubCommand: subCommand,
+	}
+	return s.Execute(command)
+}
+
+func (s *Raft) DeleteClass(name string) (uint64, error) {
+	command := &cmd.ApplyRequest{
+		Type:  cmd.ApplyRequest_TYPE_DELETE_CLASS,
+		Class: name,
+	}
+	return s.Execute(command)
+}
+
+func (s *Raft) RestoreClass(cls *models.Class, ss *sharding.State) (uint64, error) {
+	if cls == nil || cls.Class == "" {
+		return 0, fmt.Errorf("nil class or empty class name : %w", schema.ErrBadRequest)
+	}
+	req := cmd.AddClassRequest{Class: cls, State: ss}
+	subCommand, err := json.Marshal(&req)
+	if err != nil {
+		return 0, fmt.Errorf("marshal request: %w", err)
+	}
+	command := &cmd.ApplyRequest{
+		Type:       cmd.ApplyRequest_TYPE_RESTORE_CLASS,
+		Class:      cls.Class,
+		SubCommand: subCommand,
+	}
+	return s.Execute(command)
+}
+
+func (s *Raft) AddProperty(class string, props ...*models.Property) (uint64, error) {
+	for _, p := range props {
+		if p == nil || p.Name == "" || class == "" {
+			return 0, fmt.Errorf("empty property or empty class name : %w", schema.ErrBadRequest)
+		}
+	}
+	req := cmd.AddPropertyRequest{Properties: props}
+	subCommand, err := json.Marshal(&req)
+	if err != nil {
+		return 0, fmt.Errorf("marshal request: %w", err)
+	}
+	command := &cmd.ApplyRequest{
+		Type:       cmd.ApplyRequest_TYPE_ADD_PROPERTY,
+		Class:      class,
+		SubCommand: subCommand,
+	}
+	return s.Execute(command)
+}
+
+func (s *Raft) UpdateShardStatus(class, shard, status string) (uint64, error) {
+	if class == "" || shard == "" {
+		return 0, fmt.Errorf("empty class or shard : %w", schema.ErrBadRequest)
+	}
+	req := cmd.UpdateShardStatusRequest{Class: class, Shard: shard, Status: status}
+	subCommand, err := json.Marshal(&req)
+	if err != nil {
+		return 0, fmt.Errorf("marshal request: %w", err)
+	}
+	command := &cmd.ApplyRequest{
+		Type:       cmd.ApplyRequest_TYPE_UPDATE_SHARD_STATUS,
+		Class:      req.Class,
+		SubCommand: subCommand,
+	}
+	return s.Execute(command)
+}
+
+func (s *Raft) AddTenants(class string, req *cmd.AddTenantsRequest) (uint64, error) {
+	if class == "" || req == nil {
+		return 0, fmt.Errorf("empty class name or nil request : %w", schema.ErrBadRequest)
+	}
+	subCommand, err := proto.Marshal(req)
+	if err != nil {
+		return 0, fmt.Errorf("marshal request: %w", err)
+	}
+	command := &cmd.ApplyRequest{
+		Type:       cmd.ApplyRequest_TYPE_ADD_TENANT,
+		Class:      class,
+		SubCommand: subCommand,
+	}
+	return s.Execute(command)
+}
+
+func (s *Raft) UpdateTenants(class string, req *cmd.UpdateTenantsRequest) (uint64, error) {
+	if class == "" || req == nil {
+		return 0, fmt.Errorf("empty class name or nil request : %w", schema.ErrBadRequest)
+	}
+	subCommand, err := proto.Marshal(req)
+	if err != nil {
+		return 0, fmt.Errorf("marshal request: %w", err)
+	}
+	command := &cmd.ApplyRequest{
+		Type:       cmd.ApplyRequest_TYPE_UPDATE_TENANT,
+		Class:      class,
+		SubCommand: subCommand,
+	}
+	return s.Execute(command)
+}
+
+func (s *Raft) DeleteTenants(class string, req *cmd.DeleteTenantsRequest) (uint64, error) {
+	if class == "" || req == nil {
+		return 0, fmt.Errorf("empty class name or nil request : %w", schema.ErrBadRequest)
+	}
+	subCommand, err := proto.Marshal(req)
+	if err != nil {
+		return 0, fmt.Errorf("marshal request: %w", err)
+	}
+	command := &cmd.ApplyRequest{
+		Type:       cmd.ApplyRequest_TYPE_DELETE_TENANT,
+		Class:      class,
+		SubCommand: subCommand,
+	}
+	return s.Execute(command)
+}
+
+func (s *Raft) StoreSchemaV1() error {
+	command := &cmd.ApplyRequest{
+		Type: cmd.ApplyRequest_TYPE_STORE_SCHEMA_V1,
+	}
+	_, err := s.Execute(command)
+	return err
+}
+
+func (s *Raft) Execute(req *cmd.ApplyRequest) (uint64, error) {
+	t := prometheus.NewTimer(
+		monitoring.GetMetrics().SchemaWrites.WithLabelValues(
+			req.Type.String(),
+		))
+	defer t.ObserveDuration()
+
+	if s.store.IsLeader() {
+		return s.store.Execute(req)
+	}
+	if cmd.ApplyRequest_Type_name[int32(req.Type.Number())] == "" {
+		return 0, types.ErrUnknownCommand
+	}
+
+	leader := s.store.Leader()
+	if leader == "" {
+		return 0, s.leaderErr()
+	}
+	resp, err := s.cl.Apply(leader, req)
+	if err != nil {
+		return 0, err
+	}
+
+	return resp.Version, err
+}
