+++ usecases/schema/handler.go
@@ -19,7 +19,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 	command "github.com/weaviate/weaviate/cluster/proto/api"
-	"github.com/weaviate/weaviate/cluster/store"
+	clusterSchema "github.com/weaviate/weaviate/cluster/schema"
 	"github.com/weaviate/weaviate/entities/models"
 	"github.com/weaviate/weaviate/entities/schema"
 	schemaConfig "github.com/weaviate/weaviate/entities/schema/config"
@@ -66,7 +66,7 @@ type metaReader interface {
 	ClassEqual(name string) string
 	// MultiTenancy checks for multi-tenancy support
 	MultiTenancy(class string) models.MultiTenancyConfig
-	ClassInfo(class string) (ci store.ClassInfo)
+	ClassInfo(class string) (ci clusterSchema.ClassInfo)
 	// ReadOnlyClass return class model.
 	ReadOnlyClass(name string) *models.Class
 	ReadOnlySchema() models.Schema
@@ -78,7 +78,7 @@ type metaReader interface {
 	GetShardsStatus(class, tenant string) (models.ShardStatusList, error)
 
 	// WithVersion endpoints return the data with the schema version
-	ClassInfoWithVersion(ctx context.Context, class string, version uint64) (store.ClassInfo, error)
+	ClassInfoWithVersion(ctx context.Context, class string, version uint64) (clusterSchema.ClassInfo, error)
 	MultiTenancyWithVersion(ctx context.Context, class string, version uint64) (models.MultiTenancyConfig, error)
 	ReadOnlyClassWithVersion(ctx context.Context, class string, version uint64) (*models.Class, error)
 	ShardOwnerWithVersion(ctx context.Context, lass, shard string, version uint64) (string, error)
